pyproject.toml
[tool.poetry]
name = "saop-cli"
version = "0.1.0"
description = "A command-line tool for scaffolding SAOP agents."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
httpx = ">=0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
saop = "saop.cli:main"

[[tool.poetry.packages]]
include = "saop"
from = "."

[tool.uv.workspace]
members = [
"mcp/mcp-server-demo",
"mcp-server-demo",
"mcp-server",
]

[project]
name = "saop-cli"
version = "0.1.0"
description = "A command-line tool for scaffolding SAOP agents."
authors = [
{ name = "dwayne", email = "dwaynetoler@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
"python-dotenv>=1.0.0",
"fastapi>=0.111.0",
"uvicorn>=0.30.1",
"requests>=2.32.3",
"langchain-core>=0.2.14",
"langgraph>=0.0.86",
"langchain-openai>=0.1.18",
"langchain-community>=0.2.7",
"fastmcp>=0.0.1",
"aiohttp>=3.9.0",
"mcp[cli]>=1.13.0",
"a2a-sdk[grpc,http-server,telemetry]>=0.3.1",
"langchain-mcp-adapters (>=0.1.9,<0.2.0)",
"google-adk>=1.12.0",
"opentelemetry-sdk",
"opentelemetry-exporter-otlp",
"opentelemetry-instrumentation-asgi",
"python-jose[cryptography]>=3.3.0",
"passlib[bcrypt]>=1.7.4",
"python-multipart>=0.0.6",
"jinja2>=3.1.0"
]

[project.optional-dependencies]
dev = [
"pytest>=7.0",
"httpx>=0.27.0"
]